Added back the configuration option to enable or disable package overwrite. This is set per feed.

+++ NuFridge.Shared/Web/Actions/NuGetApiV2/DeletePackageAction.cs
@@ -63,7 +63,10 @@ namespace NuFridge.Shared.Web.Actions.NuGetApiV2
                 return response;
             }
 
-            packageRepository.RemovePackage(package);
+            if (package.Listed)
+            {
+                packageRepository.RemovePackage(package);
+            }
 
             return new Response { StatusCode = HttpStatusCode.OK };
         }
